drivers/net: fix number of segment storage type
Fixes: 97cb466d65c9 ("mbuf: use 2 bytes for port and nb segments")
Cc: stable@dpdk.org

+++ drivers/net/ark/ark_ethdev_rx.c
@@ -356,7 +356,7 @@ eth_ark_rx_jumbo(struct ark_rx_queue *queue,
 
 	uint16_t remaining;
 	uint16_t data_len;
-	uint8_t segments;
+	uint16_t segments;
 
 	/* first buf populated by called */
 	mbuf_prev = mbuf0;
